refactor: move RichText to common folder
