+++ src/components/Blog/Post/index.tsx
@@ -10,7 +10,7 @@ import ProgressBar from '@/components/Blog/ProgressBar'
 import BreadcrumbsNav from '@/components/Blog/BreadcrumbsNav'
 import Tags from '@/components/Blog/Tags'
 import Authors from '@/components/Blog/Authors'
-import RichText from '@/components/Campaign/RichText'
+import RichText from '@/components/common/RichText'
 import ContentsTable from '@/components/Blog/ContentsTable'
 import Socials from '@/components/Blog/Socials'
 import RelatedPosts from '@/components/Blog/RelatedPosts'
