+++ src/components/Campaign/Hero/index.tsx
@@ -4,7 +4,7 @@ import type { TypeHeroSkeleton } from '@/contentful/types'
 import type { Entry } from 'contentful'
 import { isAsset, isEntryTypeButton } from '@/lib/typeGuards'
 import layoutCss from '@/components/common/styles.module.css'
-import RichText from '@/components/Campaign/RichText'
+import RichText from '@/components/common/RichText'
 import { createImageData } from '@/lib/createImageData'
 import { SOCIAL_LOGIN_EVENTS } from '@/services/analytics/events/socialLogin'
 import { trackEvent } from '@/services/analytics/trackEvent'
