+++ src/components/common/RichText/index.tsx
@@ -14,7 +14,7 @@ import { Typography } from '@mui/material'
 import { isText } from '@/lib/typeGuards'
 import kebabCase from 'lodash/kebabCase'
 import { isTwitterUrl, isYouTubeUrl } from '@/lib/urlPatterns'
-import YouTube from '@/components/Blog/Youtube'
+import YouTube from '@/components/Blog/YouTube'
 import Twitter from '@/components/Blog/Twitter'
 
 const options: Options = {
